Move build_default_config to core.py, since that's the only place it is used
